+++ frontend/src/components/dashboard-components/DashboardControllerButtons.tsx
@@ -1,6 +1,7 @@
-import { PlayCircleOutlined, Settings, Source } from "@mui/icons-material";
+import { Home, PlayCircleOutlined, Settings, Source } from "@mui/icons-material";
 import { List, ListItem, ListItemIcon, ListItemText } from "@mui/material";
 import React from "react";
+import { useHistory } from "react-router-dom";
 import { SubPage, useSubPage } from "../../services/selectedDashboardSubPage";
 
 export default function DashboardControllerButtons() {
@@ -28,6 +29,7 @@ export default function DashboardControllerButtons() {
     dispatch(id);
   };
 
+  const history = useHistory();
   return (
     <List sx={{ marginX: { lg: 6 } }}>
       <ListItem
@@ -78,6 +80,29 @@ export default function DashboardControllerButtons() {
           />
         </ListItem>
       ))}
+      <ListItem
+          sx={{
+            display: {xs: 'flex', md: 'none'},
+            paddingX: { lg: 3 },
+            paddingY: { lg: 2 },
+            marginY: { lg: 1 },
+            borderRadius: "2px",
+          }}
+          button
+          key={99}
+          onClick={() => history.push("/")}
+        >
+          <ListItemIcon sx={{ paddingLeft: { lg: 0, xl: 2 } }}>
+            <Home />
+          </ListItemIcon>
+          <ListItemText
+            sx={{ marginLeft: { lg: 0, xl: 8 } }}
+            primary="Home page"
+            primaryTypographyProps={{
+              fontSize: { lg: "12pt", xl: "15pt" },
+            }}
+          />
+        </ListItem>
     </List>
   );
 }
