+++ frontend/src/components/dashboard-components/DashboardLeftContent.tsx
