+++ frontend/src/components/dashboard-components/DashboardRightContent.tsx
