+++ frontend/src/components/dashboard-components/ProjectBrowser.tsx
