+++ frontend/src/components/dashboard-components/ProjectRow.tsx
